+++ src/common-components/Box/Box.const.ts
@@ -1,0 +1,2 @@
+export const BOX_DEFAULT_AS = 'span' as const;
+export const BOX_CLASS = 'Ink-Box';