+++ src/common-components/Canvas/Canvas.const.ts
@@ -1,0 +1,1 @@
+export const CANVAS_CLASS = 'Ink-Canvas';