+++ src/common-components/ContextMenu/ContextMenu.const.ts
@@ -1,0 +1,4 @@
+export const CONTEXT_MENU_CLASS = 'Ink-ContextMenu';
+export const CONTEXT_MENU_ITEM_CLASS = 'Ink-ContextMenu__item';
+export const CONTEXT_MENU_ROLE = 'menu';
+export const CONTEXT_MENU_ITEM_ROLE = 'menuitem';