+++ src/common-components/Field/Field.const.ts
@@ -1,0 +1,2 @@
+export const FIELD_CLASS = 'Ink-Field';
+export const FIELD_LABEL_CLASS = 'Ink-Field__label';