+++ src/common-components/icons/Icon.const.ts
@@ -1,0 +1,2 @@
+export const ICON_CLASS = 'Ink-Icon';
+export const ICON_VIEWBOX = '0 0 24 24';