+++ src/common-components/icons/index.ts
@@ -1,0 +1,2 @@
+export { UndoIcon } from './UndoIcon';
+export { RedoIcon } from './RedoIcon';